Processor.css() takes an object
If it has a `to` property it's is passed to `postcss-url` and will
rewrite relative URLs to update them based on where they are being
written to. This is used by the browserify transform to ensure that the
files are updated for their new locations.

Passing an array of files is now expected to be on a `files` property.

Fixes #1