feat: revert #266 to be more like #158 (#347)
* feat: revert #266 to be more like #158

Fixes #345

Turns out that `rollup-watch` only calls `transform` for files that changed (were purged from the `cache` object), so only changes files exist in the second pass on a bundle. That's... bad.

Instead go back to remove dependencies and re-processing if they're "new", but only after the first run. Still feels kludgey but seems to work ok after cleaning up the watch test to be a bit more interesting.

* fix: moved specimen but left old reference

* test: ugh ugh ugh