feat: add "during" hook & support export plugins
`after` should probably just become what `during` is, it's a breaking change but not terribly different in practice.

Plugins run as part of the `during` phase can add messages with a plugin name starting with `modular-css-export` and an `exports` key that will be merged into the file's exports.

Fixes #401 in a roundabout way by making it possible w/o a change to core.

This also makes the output module support non-array values because having to know that is just an awkward implementation detail that makes no sense for anyone but me.