refactor: safer removal of changed dependencies
Instead of blindly removing all dependencies of the changed file (including all of *their* dependents) now the code will check the dependencies first and only remove ones that actually changed. This speeds things up some and also solves a few issues we were seeing with files that were both in dependency chains AND used directly by another JS file.

I'll try to write a test for that later, it won't be simple.