feat: rework svelte as class, disable default removal
Svelte preprocessor is now a class to simplify invoking functionality. Removed default cleaning because in our current most-common use-case (used with `rollup-plugin-svelte` and `@modular-css/rollup`) it actually was breaking things. Left it as an option for other use-cases where that might be wanted though.

Fixes #522

BREAKING CHANGE: if you want to remove files as they're re-processed (for standalone usage, for instance), you'll now need to pass `{ clean : true }`.