Fix: Replace standalone value references in templates
Fixed an issue where the Regex over a component in svelte's
preprocessing would try to turn a single value (in this case, an @value
for something like a color) into an array-joined-string of classes.

Values don't necessitate that logic, so I've added a check to
branch based on whether or not the css key is an array.