fix: don't mutate arrays while iterating
This cropped up in a much more complicated codebase at work. Haven't nailed down an actual repro specimen yet but verified this change works there.

Issue was that the outgoing array was being iterated with a .forEach but inside of the loop .removeDependency was being called and that mutates the array. This mean that not all entries in the array would be iterated, leading to incorrect chunking results.

Fix is simple, since dependency-graph already does the right thing and cleans up dependencies when removing a node I can get rid of all my manual cleanup code and just call .removeNode when done iterating.